+++ sample/src/main/java/tech/thdev/sample/presenter/MainContract.java
@@ -1,0 +1,19 @@
+package tech.thdev.sample.presenter;
+
+import tech.thdev.base.presenter.BasePresenter;
+import tech.thdev.base.presenter.BaseView;
+
+/**
+ * Created by Tae-hwan on 9/1/16.
+ */
+
+public interface MainContract {
+
+    interface View extends BaseView<Presenter> {
+
+    }
+
+    interface Presenter extends BasePresenter<View> {
+
+    }
+}