+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Change Log
 
+### Version 1.0.1
+
+- Change ActivityUtil
+  - 'setContentFragment' -> 'replaceContentFragment' - android.app.Fragment
+  - add 'replaceContentFragment' - import android.support.v4.app.Fragment
+                                   import android.support.v7.app.AppCompatActivity
+
 ### Version 1.0.0 (2016-09-01)
 
 - Initial release.